CartService/requirements.txt
Flask>=3.0
python-dotenv
mysql-connector-python
SQLAlchemy>=2.0
marshmallow>=3.20
requests>=2.31